Implemented reflection and refraction

+++ Raytracer/Helpers.cpp
@@ -1,3 +1,6 @@
 #include "Helpers.hpp"
 
-bool is_equal(float a, float b) { return (std::abs(a - b) < EPSILON); }
+bool is_equal(double a, double b) 
+{ 
+	return (std::abs(a - b) < EPSILON); 
+}
